+++ AIQKit/include/iQKit/iQScannerViewController.h
@@ -6,11 +6,13 @@
 
 #import "iQViewController.h"
 #import "iQScannerOverlayView.h"
+#import "iQScannerViewController.h"
 #import "iQNothingScannedViewController.h"
 #import <AVFoundation/AVFoundation.h>
 
 @class iQScannerViewController;
 @class iQAPISearchResponse;
+@class AIQScanAPIResponse;
 
 typedef NS_ENUM(NSInteger, iQScannerViewControllerIntent) {
     iQScannerViewControllerIntentBarcode,
@@ -21,9 +23,10 @@ typedef NS_ENUM(NSInteger, iQScannerViewControllerIntent) {
 
 @protocol iQScannerViewControllerDelegate <NSObject, iQViewControllerDelegate>
 
-- (void)scannerViewController:(iQScannerViewController *)scannerViewController didLoadSearchResponse:(iQAPISearchResponse *)searchResponse;
 - (void)scannerViewController:(iQScannerViewController *)scannerViewController didSearchWithKeyword:(NSString *)keyword;
 @optional
+- (void)scannerViewController:(iQScannerViewController *)scannerViewController didLoadScanResponse:(AIQScanAPIResponse *)searchResponse;
+//- (void)scannerViewController:(iQScannerViewController *)scannerViewController didLoadSearchResponse:(iQAPISearchResponse *)searchResponse;
 - (void)scannerViewController:(iQScannerViewController *)scannerViewController didScanQRCodeWithURL:(NSURL *)url;
 - (void)scannerViewController:(iQScannerViewController *)scannerViewController didTakePhoto:(NSData *)imageData;
 - (void)scannerViewController:(iQScannerViewController *)scannerViewController didPassBackFrame:(NSData *)imageData;
@@ -49,24 +52,28 @@ typedef NS_ENUM(NSInteger, iQScannerViewControllerIntent) {
  @param passBackImageData - If YES, the scanner will return the image data via the scannerViewController:didTakePhoto: delegate, and will not send a search request. Defaults to NO.
  */
 @property (nonatomic, assign) BOOL passBackImageData;
+
+/*!
+ @abstract Determines if the scanner should handle QRCode and displays QR URL.
+ 
+ @param handlesQRCode - If YES, the scanner will display a QR URL in AIQ's Webview controller, if a delegate is assisgned and implements the scannerViewController:didScanQRCodeWithURL method, AIQ's webview will not be invoked.
+ */
+@property (nonatomic, assign) BOOL handlesQRCode;
+
 @property (nonatomic, assign) BOOL passBackNextFrame;
-
 @property (nonatomic, nonatomic) BOOL continuousScanEnabled;
-
 @property (strong, nonatomic) UIView *previewView;
 @property (strong, nonatomic) UIImageView *imageView;
 @property (nonatomic, strong) iQScannerOverlayView *scannerOverlayView;
-
+@property (nonatomic, assign) BOOL hintsDisplayed;
 @property (nonatomic, strong) UIButton *menuButton;
--(void)setLeftMenuButton:(UIButton*)leftButton;
 
+-(void)setLeftMenuButton:(UIButton*)leftButton;
 -(void)resetCaptureScreen;
 -(void)startBarAnimation;
 -(void)stopScreenCapturing;
-
+-(void)setCollectionId:(NSString*)collectionId;
 - (void)showScannerOverlayView;
-
-@property (nonatomic, assign) BOOL hintsDisplayed;
 
 - (void)takePhoto;
 - (void)takePhotoWithCompletionHandler:(void (^)(NSData *imageData))handler;
@@ -76,11 +83,6 @@ typedef NS_ENUM(NSInteger, iQScannerViewControllerIntent) {
 
 -(void)startRunning;
 -(void)stopRunning;
--(void)startOpenGL;
--(void)stopOpenGL;
-
--(void)stopTimeoutDetectionTimer;
--(void)resetTimeoutDetectiontimer;
 
 - (void)setTorchMode:(AVCaptureTorchMode)torchMode;