+++ CHANGELOG.md
@@ -2,13 +2,20 @@
 All notable changes to AIQKit will be documented in this file.
 `AIQKit` adheres to [Semantic Versioning](http://semver.org/).
 --- 
+
+## [1.0.1](https://github.com/aiqtech/aiqkit-ios/releases/tag/0.9.10) (22/01/2019)
+Released on Thurs, Jul 4, 2019. 
+#### Updated
+* V1.0.1 release for AIQKit
+* Update iQKitResources.bundle with missing UI assets
+* Update and add handlesQRCode for iQScannerViewController for auto handling and display of QRCode url with iQWebviewController
+
 ## [1.0.0](https://github.com/aiqtech/aiqkit-ios/releases/tag/0.9.10) (22/01/2019)
 Released on Wed, Jul 3, 2019. 
 #### Updated
 * V1.0.0 release for AIQKit
 * Uses AIQApi cocoapod as dependency and new AIQ Platform API
 * Updated AIQ Scanner view
-
 
 ## [0.9.11](https://github.com/aiqtech/aiqkit-ios/releases/tag/0.9.10) (22/01/2019)
 Released on Tue, Jan 29, 2019.